src/requirements.txt
asgiref==3.8.1
boto3==1.35.24
botocore==1.35.24
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-allauth==64.2.1
django-analytical==3.1.0
django-appconf==1.0.6
django-axes==6.5.1
django-compressor==4.5.1
django-debug-toolbar==4.4.6
django-meta==2.4.2
django-storages==1.14.4
docutils==0.21.2
idna==3.10
jmespath==1.0.1
mypy==1.11.2
mypy-extensions==1.0.0
pilkit==3.0
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
rcssmin==1.1.2
requests==2.32.3
rjsmin==1.2.2
s3transfer==0.10.2
sentry-sdk==2.14.0
six==1.16.0
sqlparse==0.5.1
stripe==10.11.0
tornado==6.4.1
typing_extensions==4.12.2
urllib3==2.2.3
